Observer.map: Rewrite for simplicity
There were a couple problems here, first, it assumes too many things
about what the caller wants. Consolidating listeners like this is tricky
because of the calling conventions of destam, we were con correctly
implementing that here.

The old implementation was also heavy on the memory. There are circumstances
where being fancy with how we generate events actually hurts performance
through memory usage and the GC.

If users want these mechanisms, we already had a tool for that: Observer.memo