util: Look for element to remove backwards
This allows for a constant time removal with splice. With indexOf, we
enter a little bit of a dilemma: If we try to remove the first item,
indexOf will be constant time, but splice will be O(n). If we remove the
last item, we get O(n) as well but flipped. With lastIndexOf, we can achieve
O(1) upon removal of the last element. This allows for faster removal for
a common case.