Observer: Rewrite to use prototypes
Rewrite observer creation to instead rely on prototypes. Creating three
functions with a lexical environment is really expensive and really
memory heavy. By using prototypes and storing state on the observer
object, memory is cut by 50%-70%. This invokes the GC less and of the times
it is invoked, more are minor GCs. This dramatically also improves runtime.

Breaking: Observer.get/set/register_ must be called in reference to the observer